+++ src/domain/entities/Car.js
@@ -1,5 +1,5 @@
-import CONDITIONS from '../domain/constants/Conditions.js';
-import ERRORS from '../domain/constants/Errors.js';
+import CONDITIONS from '../constants/Conditions.js';
+import ERRORS from './../constants/Errors.js';
 
 class Car {
   #name;
